chore: move link metadata repository to post feature
